controller: Defer until sb data is available.
When starting the ovn-controller without monitor_all we need a few
iterations of the engine node and updating monitoring conditions until
we have loaded all data.
Previously this was handled by the number of main loop iterations of
ovn-controller and an arbitrary timeout.
However this is was always best effort and there was no guarantee that
we actually had loaded all data.

To solve this we now track the monitoring requests we sent out to
southbound and when that data becomes available.

For monitor_all we just need to have the initial monitoring conditions
fullfilled and are afterwards sure that we have all required data.

For non monitor_all we need at least one more iteration of the engine
node and updating sb monitoring conditions. We limit this to 20 rounds
for now, which should be sufficient and is at least the value of what
the previous implementation provided.
However if we notice that we do not update monitoring condition anymore
after a engine run we can also assume that we have finished loading all
data.

Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>