tests: Update ovn-sbctl count-flows test to test with NB UUIDs.
`ovn-sbctl lflow-list` and `ovn-sbctl count-lflows` can accept the UUID
of a northbound logical datapath as its argument. We didn't have
anything that tested that this works properly, so this fills the gap by
extending the existing count-flows test. We ensure that when we count
the logical switch's flows using the logical switch UUID, we get the
same results as we did when we counted using the name.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>