controller: Modify the behavior of lport_pb_is_chassis_resident.
The function behavior would be unexpected in cases when there is
BFD flap between two gateway chassis. This would result in two
chassis racing to claim the same port binding. This leads to
multiple problems like unsent gARPs. The gARP problem on its
own was solved [0], but unfortunately reintroduced [1] in
slightly different form.

The conclusion from the investigation is that the function in
the current form is easy to misuse. To prevent that use
only SB state to determine which chassis has actually claimed
the port binding. This is the desired behavior in cases where
this function was used previously.

The code that determines which chassis should claim port binding
based on BFD status remains unchanged, thus the behavior during
failover should be the same with minimizing the impact for BFD
flapping cases.

[0] 289ec19b01ad ("pinctrl: Fix missing garp.")
[1] 05527bd6ccdb ("controller: Extract garp_rarp to engine node.").

Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
(cherry picked from commit aa088003406662340df41aaac6e68e5d96272fa6)