controller: Don't run garp_rarp recompute or unrelated changes.
The garp_rarp node was doing recompute on datapaths that are not
local, this is problematic with ovn-monitor-all enabled as the
whole cluster gets notifications about unrelated port bindings.
This could lead to the node being cancelled on large scale
deployments as there is chance that SB might be readonly when
the port binding recompute is triggered.

To avoid that skip datapaths that are not local. This should prevent
an issue when the port is remote and becomes local as there are
only 2 conditions when that could happen, the datapath is already
local in that case it will be processed by the port binding handler.
In the second case the datapath is remote and becomes local when
the port is bound to local chassis, in that case we care about the
localnet.

It follows the same semantics in the opposite case local->remote,
with one exception, currently we are not removing datapath that
is local during incremental processing.

Fixes: 05527bd6ccdb ("controller: Extract garp_rarp to engine node.")
Reported-by: Ilya Maximets <i.maximets@ovn.org>
Tested-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>