ovn-northd.at: Fix unreliable 'sb_cfg propagation' test.
The test assumed the first transaction to be logged by ovn-northd is
the one corresponding to the switch datapath creation.  In some cases
however [0], likely on slower systems, the first transaction through
which ovn-northd synchronizes the initial state (e.g., RBAC) was also
happening after jsonrpc logging was enabled by the test.

Wait for a full SB sync before running the actual test.

[0] https://github.com/dceara/ovn/actions/runs/15995705960/job/45118770130#step:12:5400

Fixes: e1495fbc86bb ("northd: Remove delay in sb_cfg propagation.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>