multinode tests: Dump logs when test exits.
When a multinode test was failing, there were little or no logs
available.
Logs are added whether the test succeeds or not.
Except if requested (by -d option), test result folder for a test is
anyway removed if the test succeeds, so this patch will not dramatically
increase the size/number of logs.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>