utilities: Remove the docker directory.
The docker directory seems to be unmaintained for a while. The last
update happened in 2020 with one exception in 2024 [0]. As stated in
the commit:

"Most of the steps were inaccurate. Instead, use latest Ubuntu, use
OVS from the submodule inside the OVN repo."

Which indicates that the tooling wasn't used for a while, remove
it completely as there are other ways how to set up ovn inside
container.

[0] 77a24fc1f867 ("utilities/docker: Fix up container build.")
Acked-by: Aliasgar Ginwala <aginwala@ebay.com<mailto:aginwala@ebay.com>>
Signed-off-by: Ales Musil <amusil@redhat.com>