logical-fields: Fix IPv6 dp flow explosion caused by ip6.mcast_rsvd.
OVN routers are configured to drop any traffic with a destination
being one of the Reserved Multicast Addresses (RFC 4291).

This is done by matching on all the bits of ipv6.dst, except for bits
112-116 that cover all the addresses.  Once installed into OVS, this
turns into a following match:

  ipv6_dst=ff00::/fff0:ffff:ffff:ffff:ffff:ffff:ffff:ffff

We fixed a large chunk of IPv6 datapath flow explosion issues by
turning on prefix tacking in the flow classifier in OVS in commit
89e43f7528b0 ("controller: Fix IPv6 dp flow explosion by setting flow
table prefixes.").  However, prefix tracking doesn't work for masks
that are not contiguous.  That means that if a packet reaches a
classifier subtable with non-contiguous mask, all the bits of that
mask will be un-wildcarded.  It's not a huge problem in a general case,
because most non-contiguous masks would typically match on just a few
bits.  But ip6.mcast_rsvd is matching on 124 bits, un-wildcarding them
for most of the IPv6 traffic traversing a router and causing creation
of a separate exact-match datapath flow per destination IP.

For setups that handle large amount of traffic from many different
external addresses this issue makes IPv6 handling significantly harder
than IPv4, causing much higher load on the datapath with potential
overflow of datapath flow tables and a subsequent upcall storm.
Even without the overflow, OVS spends a lot of time revalidating all
these datapath flows burning CPU cycles.

In general, since the number of external IP addresses is virtually
unlimited, there should be no configuration where OVN exact-matches
them, otherwise it will be a significant datapath scaling issue.

Fix that by replacing a non-contiguous bit-match with a match on an
address set where all the reserved multicast addresses are just listed
directly.  There are only 16 of them, so this should not be a huge
problem to have extra 15 OpenFlow rules per router, but it will allow
OVS to use prefix tracking for these flows and avoid creating separate
datapath flow per destination IP.

Also adding a simple lsp-to-external routing test case to make sure
we don't have exact matches in this simple common use case.

The OVS classifier can likely be improved to handle non-contiguous
masks better, but it's not how the prefix tracking is designed, so
it's not a simple task.

Fixes: 677a3ba4d66b ("ovn: Add MLD support.")
Reported-at: https://issues.redhat.com/browse/FDP-1557
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Han Zhou <hzhou@ovn.org>