controller: Update next_cfg when transaction succeeded.
next_cfg updates are used as a way to know if transaction with sb got poperly committed.
Before this patch, next_cfg was only updated if ovsdb_idl_loop_commit_and_wait returned
TXN_SUCCESS, i.e. if there was not new transaction in fly.

We might have had the following case:
- Loop 1: - A transaction is sent; ovsdb_idl_loop_commit_and_wait return -1 (in_progress).
- Loop 2: - The transaction completes (in ovsdb_idl_loop_run), and a new transaction is sent.
          - ovsdb_idl_loop_commit_and_wait return -1 (new txn is in_progress).
          - next_cfg was not updated despite the fact that the first transaction succeeded.

Note that next_cfg is (still) updated when there is no commit in fly, and no transaction occurs
in the loop.

Fixes: a6b0007918f4 ("controller: Remove only commited virtual port binding requests.")
Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>