controller: Fix potentially missing vif entries in fdb.
vif1(hv1) - ls - localnet - ls - vif2(hv2)

In this configuration, with one logical switch with two vif (vif1 on
hv1 and vif2 on hv2), and a localnet port (localnet), traffic from vif1 to
vif2 goes through localnet. A vif entry update in fdb was lost if the sb
transaction was failing.

A transaction could fail in the following scenario:
- When packet from vif1 is received by br-int on hv1, vif1_mac is inserted in
  FDB with port=vif1 [1].
- When the packet is received on hv2 by localnet, if it is received before the
  FDB update [1], then vif1_mac is inserted in FDB with port=localnet [2].
  This transaction, when sent, will fail [A]. The transaction might be sent
  later (in some time) as sent by ovn-controller (while packet was forwarded
  by OVS w/o OVN intervention).
- When vif2 replies (e.g. ARP reply, echo reply), vif2_mac is inserted in
  FDB [3] with port=vif2. If the reply packet is handled in hv2 before FDB
  update [2] is sent, both updates [2] and [3] are sent at the same time.
  Hence, if the transaction fails [A], vif2_mac is not inserted in FDB.
- Finally, when reply hits hv1, if vif2_mac is not in FDB, an entry for
  vif2_mac will be inserted in FDB with port=localnet.

Hence, at the end of the game, the FDB entry for vif2_mac is localnet, and
vif2 stops receiving packets until it sends one (and updates FDB with
port=vif2).

This patch fixes this by resubmitting FDB updates for VIF entries if the
transaction failed.

Reported-at: https://issues.redhat.com/browse/FDP-1341
Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>