northd: Update virtual port on parent port update.
Previously, if there was a logical switch with a virtual port that had a
virtual parent port configured, and then the parent port was deleted and
recreated, the parent port would not be reclaimed by the chassis and
thus would not behave as the parent port, breaking traffic.

This is because virtual ports were not being updated if their parents
were recently created or deleted.

So long as virtual ports have a parent port configured, there should be
no issue deleting and recreating the parent. It should continue working
as the parent as soon as it's up again.

With this change, the parent port will function as the parent even after
deletion and recreation. The chassis will properly reclaim the virtual
port.

Fixes: b337750e45be ("northd: Incremental processing of VIF changes in 'northd' node.")
Reported-at: https://issues.redhat.com/browse/FDP-710
Co-authored-by: Mohammad Heib <mheib@redhat.com>
Signed-off-by: Mohammad Heib <mheib@redhat.com>
Signed-off-by: Rosemarie O'Riorden <rosemarie@redhat.com>
Acked-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Ales Musil <amusil@redhat.com>