controller/route-exchange: Fix route-exchange retry on error mechanism.
As pointed out in the bug report, ovn-controller was not scheduling an
immediate wake when it had detected a NetLink error due to a route
operation it tried to do.  Instead it was doing that too late, in the
en_route_exchange_status I-P node callback.  That one however is called
_before_ en_route_exchange runs.

Instead schedule an immediate wake exactly when we detect the error.  In
the next iteration en_route_exchange_status will propagate the error and
trigger en_route_exchange_status to run.

Fixes: 30587c6e4794 ("controller: Add en_route_exchange_status I-P node.")
Reported-at: https://mail.openvswitch.org/pipermail/ovs-dev/2025-August/425298.html
Reported-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>