Datapath_Binding: Separate type column and sync NB.UUID to SB.
Prior to this commit, if you wanted to find the corresponding northbound
UUID for a southbound Logical Datapath, you could find it in one of two
places:

For logical switches, it was in external-ids:logical-switch.
For logical routers, it was in external-ids:logical-router.

With this commit, ovn-northd will be using the Northbound UUID of the
logical switch or router as UUID when creating the corresponding
Southbound Datapath_Binding. This way, no matter the type of the
datapath, you can find the Northbound record by UUID. And you can
find the type of the datapath without having to potentially check
multiple external-ids keys to do so.

The "type" field is going to be used pretty heavily by northd in upcoming
patches, so instead of making it an external-id, it is now a
full-fledged column on the southbound Datapath_Binding. The type
of the logical datapath is stored in a column called "type".

Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>