controller, northd: Add logical flows to use the EVPN static FDBs.
Add extra action called "get_remote_fdb()", which is almost the same
as "get_fdb()" but goes into the new OFTABLE_GET_REMOTE_FDB table.
Use the new action in the logical flows to determine if the traffic
should go towards the remote FDB. The current implementation will
give a priority to remove FDB in case there is duplicate MAC between
the SB FDB table and the remote FDB table.

In order to make the logical action that assigns to two registers
work, add default flows for the FDB and remote FDB side tables.
The packet would be dropped otherwise if it would fall through
the table without finding any FDB. This doesn't change
the behavior for non-EVPN LS, because the packet is either dropped
or will go towards the "unknown" multicast group as previously.

There is also an option that will allow CMS to set FDB preference
(local/remote), called "dynamic-routing-fdb-prefer-local" in LS
other config. It defaults to "false", which means remote has the
priority.

Reported-at: https://issues.redhat.com/browse/FDP-1387
Acked-by: Xavier Simonart <xsimonar@redhat.com>
Acked-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Ales Musil <amusil@redhat.com>