northd: Update the SB datapath reference for Static MAC binding.
Unfortunately the previous fix wasn't enough, when we call
"sbrec_static_mac_binding_table_get_for_uuid()" it will still return
row even if was marked as deleted in the current transaction. This
would cause the Static MAC binding to be removed in one txn and added
in other one, causing traffic disruptions. To properly handle that
update the datapath reference instead. This should do the update
in a single transaction instead.

Fixes: e8a0ade51cc1 ("northd: Remove stale Static MAC Bindings during SB datapath re-creation.")
Suggested-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>