release-process: Change policy for feature review during soft freeze.
The current documentation says that new features that have been posted
for review before the soft freeze date may be considered for inclusion
in the upcoming release. The issue with this policy is that when the
soft freeze date hits, many new feature reviews get posted on the same
day. This results in an unforeseen load of reviews to complete during
the soft freeze period.

This change makes it so that only features that have received review
feedback prior to the soft freeze date are considered for inclusion in
the upcoming version. This makes it less likely for patches to show up
spontaneously at the last minute, and any that do will not be considered
for inclusion. This hopefully makes soft freeze easier to deal with and
encourages features to be developed and reviewed in a timely manner.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Mark Michelson <mmichels@redhat.com>