mac-cache: Handle never hit flows correctly.
When installing a new flow then its idle_age will start ticking from 0.
We previously treated this as the flow being used and bumped the MAC_Binding
timestamp. However if the flow has just been freshly installed it was
never used and therefore should not be bumped.
This is especially important for chassis that add the flows newly
because they became part of the relevant datapath.

In case of a large enough mac_binding_age_threshold and if the datapath
is added and removed across chassis faster than the
mac_binding_age_threshold this can actually lead to MAC_Binding that are
never removed even though they are inactive.

We also need to fix the test "MAC binding aging - probing GW router
Dynamic Neigh" since it previously relied on the fact that a newly
created flow directly triggers an update. This pushed the time of the
first probe request sufficiently for the test to work. As we now no
longer do this we can at most wait for 'cooldown_period' as afterwards a
probe might be send immediately.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
Signed-off-by: Ales Musil <amusil@redhat.com>
(cherry picked from commit fbb501cc8b7dc6d72618b7d25e99ac3dc8d178dd)