northd: Always prefer bound ovn_port versions when available.
Once the set of OVN ports is computed (in memory) in ovn-northd, there's
no need to ever prefer OVN ports that are unbound (no SB datapath linked
to them) in later stages.  They're actually removed after the set of
ports is built in build_ports().

The only case in which "not bound" ports were potentially found and
returned while "bound" versions of the ports also existed was when
connecting logical switch and router peer ports together,
ovn_port_get_peer(), in build_ports() potentially causing issues (and
logical flows to be partially recomputed in subsequent iterations of
ovn-northd.

Fixes: 7b404e68b5b8 ("northd: Fix duplicate logical port detection.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>