en-learned-route-sync: Fix Learned_Route cleanup for removed router ports.
It's not enough to check that ovn_port_find() returns non-NULL, that
also happens when in the current iteration the router (and router port)
have been deleted and new ones were added (with the same router port
name).  However, the Learned_Route was learned by ovn-controller for the
old Port_Binding and is potentially not applicable anymore.

Remove it by making sure we also match on the actual
Learned_Route.logical_port Port_Binding pointer being the correct one
when determining if a dynamic route is valid.

Otherwise we end up with an inconsistent SB database transaction that
fails integrity checks on the server side.

Fixes: 0ccaea177420 ("northd: Remove learned routes if lrp is removed.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Felix Huettner <felix.huettner@stackit.cloud>