controller: Add missing monitor conditions.
We were missing monitor conditions when another chassis was
requested-chassis. The port_binding (with requested-chassis set to a
different chassis) was in that case sometimes not monitored.

The additional_chassis field was not monitored neither (and is still
not) as it is redundant with this change and would increase the size
of the SB monitor condition.

The test "Multichassis port I-P processing" sometimes highlited the
issue. It has been modified so that the condition leading to the issue
would occur much more often.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>