tests: Use localhost when setting "wrong" ovn-remote.
In some isolated environments (e.g. in nixpkgs build sandbox), the
network namespace doesn't have any routes or interfaces but `lo`. In
this case, an attempt to connect to 192.168.0.10 results in ENETUNREACH,
producing an unexpected "Network unreachable" warning message in service
log file - breaking the test cleanup checks.

Since the test case doesn't seem to care if the address is available, as
long as there is no SB database actually running at the configured
ovn-remote port, use the localhost address instead (which is always
present, even in the most isolated environments).

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>
Signed-off-by: Ales Musil <amusil@redhat.com>