northd: Fix advertised route for ECMP static route.
If logical router has ECMP for a same prefix, northd tries to
adversited N times by the same logical router port, insert in
sbrec_advertised_route table fails because all fields of
sbrec_advertised_route are equal.
To fix it, just add one time the route to be advertised in the
sync routes.

Fixes: f2deb24c5c43 ("northd: Sync Advertised_Route to sb.")
Signed-off-by: Lucas Vargas Dias <lucas.vdias@luizalabs.com>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>