northd: Fix receiving NA on non resident chassis.
Since [0], ND/NA received on non-resident chassis are dropped/ignored
on DGP to avoid having all chassis of an HA group trying to update
mac_binding, which was resulting in transaction errors.
However, it can happen that the whole pipeline, including the distributed
gateway router, is run on a chassis where the DGP is not resident.
This happens for instance in the following setup, with vif1 and ext on hv1
and lr-pub on hv2.
vif1 --- ls1 --- lr --- pub -+- ln
                             +- ext

In this setup, unicast NA from ext towards vif1 should not be ignored,
even on the non-resident chassis.

This patch amends [0] to only ignore multicast ND/NA on non-resident chassis.
[0]: 987adc9e589e ("controller: Avoid IPv6 Mac_Binding related transaction errors.")

Fixes: 987adc9e589e ("controller: Avoid IPv6 Mac_Binding related transaction errors.")

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Reported-at: https://issues.redhat.com/browse/FDP-1567
Reported-at: https://issues.redhat.com/browse/FDP-1728
Signed-off-by: Ales Musil <amusil@redhat.com>