treewide: Check for proper width instead of generic LLONG_MAX.
While the long long is defined as minimal of 64 bit width, meaning
that it could change in the future to be wider e.g. 128. However, the
*_cfg columns are defined as int64_t so always being 64 bits. Replace
the LLONG_MAX in potential overflow checks with INT64_MAX instead,
that will guarantee that even a change of long long width in the
future won't break this.

Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>