Add lr_stateful_handler to en_sync_sb_pb.
Adding sync_to_sb_pb_lr_stateful_handler() allows port_bindings to be
computed in the incremental processor instead of always fully
recalculating when the lr_stateful engine is updated.

The process is taking the updated lr_stateful_record, getting the
logical router it references and syncing the routers ports and peers to
those router ports.

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>
Acked-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>