ovn: Add SNI support to ovn-nbctl and ovn-sbctl.
This adds the case for OPT_SSL_SERVER_NAME in ovn-dbctl.c,
enabling --ssl-server-name support for ovn-nbctl and ovn-sbctl.

The --ssl-server-name option allows specifying the server name for TLS
Server Name Indication (SNI), which is useful when connecting through
proxies or service meshes.

Example usage:
ovn-nbctl --db=ssl:$ISTIO_GATEWAY_IP:6641 \
    --ssl-server-name=hv-171.ovsdb.provider.internal \
    --private-key=ovn-nbctl-client.key \
    --certificate=ovn-nbctl-client.crt \
    --ca-cert=ca.crt \
    list nb_global

The above command will connect to a istio proxy. Depending
on the configuration of istio proxy for SNI hv-171,
it can connect to a pod or a node in the kubernetes cluster.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>